nyc-311-analytics/requirements.txt
streamlit==1.29.0
langgraph==0.2.0
langchain==0.1.0
langchain-community==0.1.0
duckdb==0.9.2
pandas==2.1.4
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0
openai
